refactor(web-client,helpers): move checkClass to non-test helpers
