+++ web-client/src/helpers/helpers.spec.ts
@@ -1,4 +1,4 @@
-import { ifDefined, never } from 'src/helpers/helpers';
+import { checkClass, Constructor, ifDefined, never } from 'src/helpers/helpers';
 import { assertConsoleLogs } from 'src/tests/test.helpers';
 
 describe('never', () => {
@@ -29,5 +29,31 @@ describe('ifDefined', () => {
   it('applies to defined', () => {
     const value = 5;
     expect(ifDefined(value, (n) => n * 2)).toBe(10);
+  });
+});
+
+describe('checkClass', () => {
+  class Mock {
+    public constructor() {}
+  }
+
+  it('', () => {
+    const cases: Array<[unknown, Constructor<unknown>]> = [
+      [new Map(), Map],
+      [new Mock(), Mock],
+    ];
+    cases.forEach(([o, cls]) => {
+      expect(checkClass(o, cls)).toBe(o);
+    });
+  });
+
+  it('throws error', () => {
+    const cases: Array<[unknown, Constructor<unknown>, string]> = [
+      [new Mock(), Map, 'expected Map, got Mock: [object Object]'],
+      [new Map(), Mock, 'expected Mock, got Map: [object Map]'],
+    ];
+    cases.forEach(([o, cls, message]) => {
+      expect(() => checkClass(o, cls)).toThrowError(message);
+    });
   });
 });
